backend/requirements.txt
fastapi
uvicorn
ultralytics
torch
torchvision
sqlalchemy
python-multipart
requests
python-dotenv
pillow
